Correct dz17
