--- dz17/.eslintrc.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: '@shcherbin/eslint-config-node',
-  rules: {
-    'no-console': 'off'
-  }
-}
