+++ dz17/eslint.config.js
@@ -1,0 +1,10 @@
+import config from '@shcherbin/eslint-config'
+
+export default [
+  config.browser,
+  {
+    rules: {
+      'no-console': 'off'
+    }
+  }
+]
